ChangeDriveLetter - GUI
GUI version of my previous change drive letter project(program that uses diskpart to change the drive letter of a specified drive.)made using java. Download only the -exe folder and run the shortcut to use the program, the -source folder is if you want to make any changes to it, note that to run the project from an IDE(i used NetBeans) you have to copy the contents of the resources folder and paste them in the root of the project folder.